requirements.txt
asn1crypto==0.24.0
autopep8==1.4.4
beautifulsoup4==4.8.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
future==0.17.1
idna==2.8
pep8==1.7.1
pycodestyle==2.5.0
pycparser==2.19
pyTelegramBotAPI==3.6.6
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
steamdb-parser-mrmarble==0.0.1
urllib3==1.25.3
